fix: typos and routes

+++ src/services/bookmark-service.ts
@@ -12,6 +12,8 @@ async function createBookmark(userId: number, bookId: number) {
 	if (!userId || !bookId) throw badRequestError();
 	const book = await bookRepository.findBookById(bookId);
 	if (!book) return bookNotFoundError();
+	const existingBookmark = await bookmarkRepository.bookmarkExists(userId, bookId);
+	if (existingBookmark) throw badRequestError()
 	const bookmark = await bookmarkRepository.createBookmark(userId, bookId);
 	return bookmark;
 }
@@ -20,6 +22,8 @@ async function removeBookmark(userId: number, bookId: number) {
 	if (!userId || !bookId) throw badRequestError();
 	const book = await bookRepository.findBookById(bookId);
 	if (!book) return bookNotFoundError();
+	const existingBookmark = await bookmarkRepository.bookmarkExists(userId, bookId);
+	if (!existingBookmark) throw badRequestError();
 	const bookmark = await bookmarkRepository.removeBookmark(userId, bookId);
 	return bookmark;
 }
